+++ LoginForm/src/Employees.java
@@ -348,21 +348,25 @@ public class Employees extends javax.swing.JFrame {
 
     // When delete button is pressed, delete the selected row
     private void bDeleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_bDeleteActionPerformed
-        createConnection();
-        int rowSelected = tbEmployees.getSelectedRow();
-        String selectEmpId = tbEmployees.getValueAt(rowSelected, 0).toString();
-        System.out.println(selectEmpId);
-        //deletion from the employees according to the employee id
-        String sql = "delete from employee where employeeid = '" + selectEmpId + "'";
-        try{
-            PreparedStatement ps = conn.prepareStatement(sql);
-            rs = ps.executeQuery();
-            updateTable();
-            closeConnection();
-        }
-        catch(Exception e){
-            JOptionPane.showMessageDialog(null, "can not delete");
-            System.out.println(e);
+        if(tbEmployees.getSelectedRow() == -1)
+            JOptionPane.showMessageDialog(null, "Please select a row to delete!");
+        else{
+            createConnection();
+            int rowSelected = tbEmployees.getSelectedRow();
+            String selectEmpId = tbEmployees.getValueAt(rowSelected, 0).toString();
+            System.out.println(selectEmpId);
+            //deletion from the employees according to the employee id
+            String sql = "delete from employee where employeeid = '" + selectEmpId + "'";
+            try{
+                PreparedStatement ps = conn.prepareStatement(sql);
+                rs = ps.executeQuery();
+                updateTable();
+                closeConnection();
+            }
+            catch(Exception e){
+                JOptionPane.showMessageDialog(null, "can not delete");
+                System.out.println(e);
+            }
         }
     }//GEN-LAST:event_bDeleteActionPerformed
